+++ Sources/VComponents/Components/Buttons/Stretched Button/VStretchedButton.swift
@@ -64,14 +64,14 @@ public struct VStretchedButton<Label>: View where Label: View {
     public init(
         uiModel: VStretchedButtonUIModel = .init(),
         action: @escaping () -> Void,
-        icon: Image,
-        title: String
+        title: String,
+        icon: Image
     )
         where Label == Never
     {
         self.uiModel = uiModel
         self.action = action
-        self.label = .iconTitle(icon: icon, title: title)
+        self.label = .titleAndIcon(title: title, icon: icon)
     }
     
     /// Initializes `VStretchedButton` with action and label.
@@ -117,12 +117,21 @@ public struct VStretchedButton<Label>: View where Label: View {
             case .icon(let icon):
                 iconLabelComponent(internalState: internalState, icon: icon)
 
-            case .iconTitle(let icon, let title):
-                HStack(spacing: uiModel.iconAndTitleTextSpacing, content: {
-                    iconLabelComponent(internalState: internalState, icon: icon)
-                    titleLabelComponent(internalState: internalState, title: title)
-                })
-                
+            case .titleAndIcon(let title, let icon):
+                switch uiModel.titleAndIconPlacement {
+                case .titleAndIcon:
+                    HStack(spacing: uiModel.titleTextAndIconSpacing, content: {
+                        titleLabelComponent(internalState: internalState, title: title)
+                        iconLabelComponent(internalState: internalState, icon: icon)
+                    })
+
+                case .iconAndTitle:
+                    HStack(spacing: uiModel.titleTextAndIconSpacing, content: {
+                        iconLabelComponent(internalState: internalState, icon: icon)
+                        titleLabelComponent(internalState: internalState, title: title)
+                    })
+                }
+
             case .label(let label):
                 label(internalState)
             }
@@ -236,8 +245,8 @@ struct VStretchedButton_Previews: PreviewProvider {
 
                 VStretchedButton(
                     action: {},
-                    icon: icon,
-                    title: title
+                    title: title,
+                    icon: icon
                 )
                 .modifier(StretchedButtonWidthModifier())
                 .padding(.horizontal)
@@ -355,8 +364,8 @@ struct VStretchedButton_Previews: PreviewProvider {
                         return uiModel
                     }(),
                     action: {},
-                    icon: icon,
-                    title: title
+                    title: title,
+                    icon: icon
                 )
                 .modifier(StretchedButtonWidthModifier())
                 .padding(.horizontal)