+++ Sources/VComponents/Components/Buttons/Stretched Button/VStretchedButtonLabel.swift
@@ -12,6 +12,6 @@ import SwiftUI
 enum VStretchedButtonLabel<Label> where Label: View {
     case title(title: String)
     case icon(icon: Image)
-    case iconTitle(icon: Image, title: String)
+    case titleAndIcon(title: String, icon: Image)
     case label(label: (VStretchedButtonInternalState) -> Label)
 }