+++ Sources/VComponents/Components/Buttons/Wrapped Button/VWrappedButton.swift
@@ -62,14 +62,14 @@ public struct VWrappedButton<Label>: View where Label: View {
     public init(
         uiModel: VWrappedButtonUIModel = .init(),
         action: @escaping () -> Void,
-        icon: Image,
-        title: String
+        title: String,
+        icon: Image
     )
         where Label == Never
     {
         self.uiModel = uiModel
         self.action = action
-        self.label = .iconTitle(icon: icon, title: title)
+        self.label = .titleAndIcon(title: title, icon: icon)
     }
     
     /// Initializes `VWrappedButton` with action and label.
@@ -116,12 +116,21 @@ public struct VWrappedButton<Label>: View where Label: View {
             case .icon(let icon):
                 iconLabelComponent(internalState: internalState, icon: icon)
 
-            case .iconTitle(let icon, let title):
-                HStack(spacing: uiModel.iconAndTitleTextSpacing, content: {
-                    iconLabelComponent(internalState: internalState, icon: icon)
-                    titleLabelComponent(internalState: internalState, title: title)
-                })
-                
+            case .titleAndIcon(let title, let icon):
+                switch uiModel.titleAndIconPlacement {
+                case .titleAndIcon:
+                    HStack(spacing: uiModel.titleTextAndIconSpacing, content: {
+                        titleLabelComponent(internalState: internalState, title: title)
+                        iconLabelComponent(internalState: internalState, icon: icon)
+                    })
+
+                case .iconAndTitle:
+                    HStack(spacing: uiModel.titleTextAndIconSpacing, content: {
+                        iconLabelComponent(internalState: internalState, icon: icon)
+                        titleLabelComponent(internalState: internalState, title: title)
+                    })
+                }
+
             case .label(let label):
                 label(internalState)
             }
@@ -226,8 +235,8 @@ struct VWrappedButton_Previews: PreviewProvider {
 
                 VWrappedButton(
                     action: {},
-                    icon: icon,
-                    title: title
+                    title: title,
+                    icon: icon
                 )
             })
         }
@@ -350,8 +359,8 @@ struct VWrappedButton_Previews: PreviewProvider {
                         return uiModel
                     }(),
                     action: {},
-                    icon: icon,
-                    title: title
+                    title: title,
+                    icon: icon
                 )
             })
         }