+++ Sources/VComponents/Models/TitleAndIconPlacement.swift
@@ -1,0 +1,18 @@
+//
+//  TitleAndIconPlacement.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 13.11.23.
+//
+
+import Foundation
+
+// MARK: - Title and Icon Placement
+/// Enumeration that represents title and icon placement.
+public enum TitleAndIconPlacement {
+    /// Title and icon.
+    case titleAndIcon
+
+    /// Icon and titile.
+    case iconAndTitle
+}